finish fire

+++ Assets/Scripts/Enemy.cs
@@ -3,11 +3,11 @@ using UnityEngine;
 public class Enemy : MonoBehaviour
 {
    public int maxHealth = 100;
-   public int health = 100;
+   public float health = 100;
    
    public void TakeDamage(float damage)
    {
-      health -= (int)damage;
+      health -= damage;
       if (health <= 0)
       {
          Die();
